+++ src/main/java/be/kdg/ip3/archportal/lobbies/application/PartyService.java
@@ -13,8 +13,8 @@ import be.kdg.ip3.archportal.lobbies.domain.party.Party;
 import be.kdg.ip3.archportal.lobbies.domain.party.PartyId;
 import be.kdg.ip3.archportal.lobbies.domain.party.PartyRepository;
 import be.kdg.ip3.archportal.lobbies.domain.partyInvite.PartyInvite;
-import be.kdg.ip3.archportal.lobbies.shared.DeletedPartyEvent;
 import be.kdg.ip3.archportal.lobbies.shared.JoinedPartyEvent;
+import be.kdg.ip3.archportal.lobbies.shared.LeftPartyEvent;
 import be.kdg.ip3.archportal.profiles.shared.ProfilesApi;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -124,11 +124,8 @@ public class PartyService {
             throw playerId.notFound();
         var party = partyRepository.findByMemberId(playerId).orElseThrow(() -> new NotFoundException("Party not found"));
         party.leaveParty(playerId);
-        if (party.getHostId() == null) {
-            partyRepository.deleteById(party.getId());
-            publisher.publishEvent(new DeletedPartyEvent(party.getChatRoomId().id()));
-        } else {
-            partyRepository.save(party);
-        }
+        if (party.getHostId() == null) partyRepository.deleteById(party.getId());
+        else partyRepository.save(party);
+        publisher.publishEvent(new LeftPartyEvent(party.getChatRoomId().id(), playerId.id()));
     }
 }
